updates lambda api to refresh auth token
