+++ src/api/index.ts
@@ -1,5 +1,4 @@
 export { default } from "./newm/api";
-export * from "./newm/utils";
 export * from "./newm/types";
 export { default as cloudinaryApi } from "./cloudinary/api";
 export * from "./cloudinary/types";
